+++ src/middlewares/validateSharedTokenMiddleware.ts
